+++ example/source-embed-packages/xcore/util/time.cc
@@ -1,0 +1,11 @@
+#include "util/time.h"
+
+namespace xmake {
+namespace xcore {
+
+uint64_t Time::ToNanosecond() const {
+  return nanoseconds_;
+}
+
+}
+}