+++ example/source-embed-packages/xgpu/main.cc
@@ -1,0 +1,7 @@
+#include "math/vec2d.h"
+
+int main() {
+    xmake::xmath::Vec2d(12.34, 56.78);
+
+    std::cout << "hello" << std::endl;
+}