+++ example/source-embed-packages/xgpu/test.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+xmake build --yes --all || exit 1